+++ src/error.rs
@@ -74,6 +74,10 @@ pub(crate) enum MockError {
         pat_index: PatIndex,
         lender: Lender,
     },
+    CannotReturnValueMoreThanOnce {
+        fn_call: FnCall,
+        pat_index: PatIndex,
+    },
     FailedVerification(String),
     CannotUnmock {
         name: &'static str,
@@ -88,25 +92,25 @@ pub(crate) enum MockError {
 impl std::fmt::Display for MockError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            MockError::Downcast { name } => {
+            Self::Downcast { name } => {
                 write!(f, "Fatal: Failed to downcast for {name}.")
             }
-            MockError::NoMockImplementation { name } => {
+            Self::NoMockImplementation { name } => {
                 write!(f, "No mock implementation found for {name}.")
             }
-            MockError::NoMatchingCallPatterns { fn_call } => {
+            Self::NoMatchingCallPatterns { fn_call } => {
                 write!(f, "{fn_call}: No matching call patterns.")
             }
-            MockError::NoOutputAvailableForCallPattern {
+            Self::NoOutputAvailableForCallPattern {
                 fn_call,
                 pat_index,
             } => {
                 write!(f, "{fn_call}: No output available for matching call pattern {pat_index}.")
             }
-            MockError::MockNeverCalled { name } => {
+            Self::MockNeverCalled { name } => {
                 write!(f, "Mock for {name} was never called. Dead mocks should be removed.")
             }
-            MockError::CallOrderNotMatchedForMockFn {
+            Self::CallOrderNotMatchedForMockFn {
                 fn_call,
                 actual_call_order,
                 expected_ranges,
@@ -114,23 +118,23 @@ impl std::fmt::Display for MockError {
                 let ranges_dbg = expected_ranges.iter().map(format_call_range).collect::<Vec<_>>().join(", ");
                 write!(f, "{fn_call}: Matched in wrong order. It supported the call order ranges [{ranges_dbg}], but actual call order was {actual_call_order}.")
             }
-            MockError::InputsNotMatchedInCallOrder {
+            Self::InputsNotMatchedInCallOrder {
                 fn_call,
                 actual_call_order,
                 pat_index,
             } => {
                 write!(f, "{fn_call}: Invoked in the correct order ({actual_call_order}), but inputs didn't match call pattern {pat_index}.")
             }
-            MockError::TypeMismatchExpectedOwnedInsteadOfBorrowed {
+            Self::TypeMismatchExpectedOwnedInsteadOfBorrowed {
                 fn_call,
                 pat_index,
             } => write!(f, "{fn_call}: Type mismatch: Expected an owned return value, but found a borrow for call pattern {pat_index}. Try using Match::returns() or Match::answers()."),
-            MockError::CannotBorrowValueProducedByClosure {
+            Self::CannotBorrowValueProducedByClosure {
                 fn_call,
                 pat_index,
                 lender,
             } => write!(f, "{fn_call}: Cannot borrow the value returned by the answering closure for pattern {pat_index}. {}", lender.suggest()),
-            MockError::CannotBorrowInvalidLifetime {
+            Self::CannotBorrowInvalidLifetime {
                 fn_call,
                 pat_index,
                 lender,
@@ -139,11 +143,17 @@ impl std::fmt::Display for MockError {
                 Lender::Param =>write!(f, "{fn_call}: Cannot borrow output value from a parameter for call pattern {pat_index}. {}", lender.suggest()),
                 Lender::Static => write!(f, "{fn_call}: Cannot borrow output value statically for call pattern {pat_index}. {}", lender.suggest()),
             },
-            MockError::FailedVerification(message) => write!(f, "{message}"),
-            MockError::CannotUnmock { name } => {
+            Self::CannotReturnValueMoreThanOnce {
+                fn_call,
+                pat_index,
+            } => {
+                write!(f, "{fn_call}: Cannot return value more than once for call pattern {pat_index}, because of implicit quantification to `once()`. Try adding `.cloned()` or similar.")
+            },
+            Self::FailedVerification(message) => write!(f, "{message}"),
+            Self::CannotUnmock { name } => {
                 write!(f, "{name} cannot be unmocked as there is no function available to call.")
             },
-            MockError::ExplicitPanic { fn_call, pat_index, msg } => write!(f, "{fn_call}: Explicit panic for call pattern {pat_index}: {msg}")
+            Self::ExplicitPanic { fn_call, pat_index, msg } => write!(f, "{fn_call}: Explicit panic for call pattern {pat_index}: {msg}")
         }
     }
 }