+++ tests/test_mock_order.rs
@@ -17,9 +17,9 @@ trait T2 {
 )]
 fn two_fns_in_incorrect_order_should_fail_and_presence_of_a_stub_should_not_influence_order() {
     let m = mock([
-        T1__a.next_call(matching!(0)).returns(0).once().in_order(),
+        T1__a.next_call(matching!(0)).returns(0).in_order(),
         T1__b.stub(|each| {
-            each.call(matching!(_)).returns(0);
+            each.call(matching!(_)).returns(0).cloned();
         }),
         T2__c.next_call(matching!(0)).returns(0).once().in_order(),
         T1__a.next_call(matching!(0)).returns(0).once().in_order(),